fix(checkout): Make purchase line-item quantity an integer
Rust used f64 while Convex used v.number(); fractional quantities are not
meaningful for line items and the types disagreed. Now u32 end to end.

+++ crates/sprocket-agent/src/tools.rs
@@ -498,7 +498,7 @@ pub(crate) struct PurchaseItem {
     description: String,
     unit_price: String,
     #[serde(skip_serializing_if = "Option::is_none")]
-    quantity: Option<f64>,
+    quantity: Option<u32>,
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
